backend/requirements.txt
flask
python-dotenv
openai
pillow
python-docx